requirements.txt
# Underworld Empire — Python dependencies
discord.py>=2.3.0
motor>=3.3.0
pymongo>=4.6.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
